Add configure-namespace action
