+++ README.md
@@ -4,6 +4,8 @@
 
 # PorterOps: Porter Operator
 
+🚨 This is a new project and the goals below are aspirational and not all implemented yet.
+
 PorterOps not only gives you a native, integrated experience for managing your
 bundles with Kubernetes but is the recommended way to automate your bundle
 pipeline with support for GitOps.
@@ -18,13 +20,93 @@ pipeline with support for GitOps.
 
 <p align="center">Learn all about Porter at <a href="https://porter.sh">porter.sh</a></p>
 
+# Install
+
+Use explain to see what credentials and parameters you can use when installing and configuring the operator.
+```
+porter explain -r ghcr.io/getporter/porter-operator:canary
+```
+
+Generate a credential set for the bundle
+```
+porter credentials generate porterops -r ghcr.io/getporter/porter-operator:canary
+```
+
+Install the operator into the porter-operator-system namespace
+```
+porter install porterops -c porterops -r ghcr.io/getporter/porter-operator:canary
+```
+
+Create a namespace with the appropriate rbac, secrets and configmaps. This is where you will run porter.
+
+```
+porter invoke porterops --action configure-namespace --param namespace=TODO -c porterops
+```
+
+# Install a bundle
+
+Here is an example installation CRD:
+
+```yaml
+apiVersion: porter.sh/v1
+kind: Installation
+metadata:
+  name: porter-hello
+spec:
+  reference: "getporter/porter-hello:v0.1.1"
+  action: "install"
+```
+
+After you apply it with `kubectl apply -f`, the Porter Operator will run the following command:
+
+```
+porter install porter-hello getporter/porter-hello:v0.1.1
+```
+
+# Upgrade a bundle
+
+Edit the installation CRD and change the action to "upgrade":
+
+```yaml
+apiVersion: porter.sh/v1
+kind: Installation
+metadata:
+  name: porter-hello
+spec:
+  reference: "getporter/porter-hello:v0.1.1"
+  action: "upgrade"
+```
+
+# Retry the last operator
+
+If your bundle operation failed, you can run it again by changing an annotation on the installation CRD:
+
+```
+apiVersion: porter.sh/v1
+kind: Installation
+metadata:
+  name: porter-hello
+  annotations:
+    retry: 1
+spec:
+  reference: "getporter/porter-hello:v0.1.1"
+  action: "upgrade"
+```
+
+Each time you need to repeat the operation, change the annotation value again.
+There is nothing special about the key used for the annotation. I chose retry,
+however you could use "favorite-color: blue", changing the value each time, and
+it would still trigger Porter to retry it. 
+
+# Configure the Operator
+This section breaks down what the configure-namespace action of the bundle is
+doing under the hood.
+
 ## Define Secrets
 
-The operator uses secrets defined in namespace of the CRD being managed to populate
+The operator uses secrets defined in the namespace of the CRD being managed to populate
 Porter's config file and environment variables. For example the azure connection
 string, or service principal environment variables.
-
-TODO: Have porter help do these parts
 
 ### porter-config
 These secrets are copied into the pod as files to tell Porter where to save its data
@@ -41,11 +123,13 @@ for the plugins so that they can connect to remote services such storage or vaul
 
 ```
 kubectl create secret generic porter-env \
-  --from-literal=AZURE_STORAGE_CONNECTION_STRING=$PORTER_TEST_AZURE_STORAGE_CONNECTION_STRING \
-  --from-literal=AZURE_CLIENT_SECRET=$PORTER_AZURE_CLIENT_SECRET \
-  --from-literal=AZURE_CLIENT_ID=$PORTER_AZURE_CLIENT_ID \
-  --from-literal=AZURE_TENANT_ID=$PORTER_AZURE_TENANT_ID
+  --from-literal=AZURE_STORAGE_CONNECTION_STRING=$AZURE_STORAGE_CONNECTION_STRING \
+  --from-literal=AZURE_CLIENT_SECRET=$AZURE_CLIENT_SECRET \
+  --from-literal=AZURE_CLIENT_ID=$AZURE_CLIENT_ID \
+  --from-literal=AZURE_TENANT_ID=$AZURE_TENANT_ID
 ``` 
+
+Right now the bundle only works with azure plugin.
 
 ## Define Configuration
 
@@ -54,9 +138,14 @@ These are configuration settings for the Porter Operator.
 
 ```
 kubectl create configmap porter \
+  --from-literal=serviceAccount=porter-agent \
   --from-literal=porterVersion=canary \
-  --from-literal=serviceAccount=porter-agent
+  --from-literal=porterRepository=ghcr.io/getporter/porter-operator \
+  --from-literal=volumeSize=64Mi \
+  --from-literal=pullPolicy=Always
 ```
+
+Only serviceAccount is required, the rest can be omitted.
 
 See [Modify the porter agent](/CONTRIBUTING.md#modify-the-porter-agent) for details on 
 how this is created.
