+++ config/crd/kustomization.yaml
@@ -2,7 +2,8 @@
 # since it depends on service name and namespace that are out of this kustomize package.
 # It should be run by config/default
 resources:
-- bases/porter.sh_installations.yaml
+  - bases/porter.sh_agentconfigs.yaml
+  - bases/porter.sh_installations.yaml
 # +kubebuilder:scaffold:crdkustomizeresource
 
 patchesStrategicMerge:
