+++ installer/Dockerfile.tmpl
@@ -1,0 +1,13 @@
+FROM debian:stretch-slim
+
+ARG BUNDLE_DIR
+
+RUN apt-get update && apt-get install -y ca-certificates curl
+
+RUN curl -sLo /usr/bin/yq https://github.com/mikefarah/yq/releases/download/v4.5.1/yq_linux_amd64 && \
+    chmod +x /usr/bin/yq
+
+# PORTER_MIXINS
+
+# Use the BUNDLE_DIR build argument to copy files into the bundle
+COPY . $BUNDLE_DIR
