+++ installer/manifests/namespace/namespace.yaml
@@ -1,0 +1,6 @@
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: REPLACE
+  labels:
+    porter: "true"
