+++ installer/manifests/namespace/porter-config-secret.yaml
@@ -1,0 +1,9 @@
+apiVersion: v1
+kind: Secret
+metadata:
+  name: porter-config
+  namespace: REPLACE
+  labels:
+    porter: "true"
+data:
+  config.toml:
