+++ installer/manifests/namespace/porter-config/.keep
@@ -1,0 +1,1 @@
+Files in this directory will be included to the porter configmap
