+++ installer/manifests/namespace/porter-configmap.yaml
@@ -1,0 +1,8 @@
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: porter
+  namespace: REPLACE
+  labels:
+    porter: "true"
+data:
