+++ installer/manifests/namespace/porter-env-secret.yaml
@@ -1,0 +1,8 @@
+apiVersion: v1
+kind: Secret
+metadata:
+  name: porter-env
+  namespace: REPLACE
+  labels:
+    porter: "true"
+stringData:
